Restructure Session to follow new paradigm
